refactor: extract BuilderFields into separate file
